Save as starting point for refactoring.
Refactoring test case:
build underground with intact bricks; the repair bricks are spawned only in the first
column in succession (0,0), (0,1), (0,2) first with state Broken and then with state
Destroyed; repair stones only in the first column. As a result, all ground bricks in
the end have the state intact, the repair bricks are stacked with State Repair.